refactor/authenticator: move app auth routines to a separate module
Also, implement the bare-bones ops recovery testing framework
which allows to override responses for certain requests in mock routing
