+++ safe_authenticator/src/lib.rs
@@ -69,6 +69,7 @@ pub mod ipc;
 pub mod public_id;
 
 mod access_container;
+mod app_auth;
 mod app_container;
 mod config;
 mod errors;
@@ -88,14 +89,16 @@ use futures::sync::mpsc;
 use maidsafe_utilities::serialisation::serialise;
 use maidsafe_utilities::thread::{self, Joiner};
 use routing::EntryActions;
-use safe_core::{Client, CoreMsg, CoreMsgTx, FutureExt, MDataInfo, NetworkEvent, event_loop,
-                mdata_info};
+use safe_core::{Client, CoreError, CoreMsg, CoreMsgTx, FutureExt, MDataInfo, NetworkEvent,
+                NetworkTx, event_loop, mdata_info};
+#[cfg(feature = "use-mock-routing")]
+use safe_core::MockRouting;
 use safe_core::ipc::Permission;
 use safe_core::nfs::{create_dir, create_std_dirs};
 use std::collections::{BTreeSet, HashMap};
 use std::sync::Mutex;
 use std::sync::mpsc::sync_channel;
-use tokio_core::reactor::Core;
+use tokio_core::reactor::{Core, Handle};
 
 /// Future type specialised with `AuthError` as an error type
 pub type AuthFuture<T> = Future<Item = T, Error = AuthError>;
@@ -135,17 +138,34 @@ impl Authenticator {
         locator: S,
         password: S,
         invitation: S,
-        mut network_observer: NetObs,
+        network_observer: NetObs,
     ) -> Result<Self, AuthError>
     where
-        S: Into<String>,
         NetObs: FnMut(Result<NetworkEvent, ()>) + Send + 'static,
+        S: Into<String>,
     {
-        let (tx, rx) = sync_channel(0);
-
         let locator = locator.into();
         let password = password.into();
         let invitation = invitation.into();
+
+        Self::create_acc_impl(
+            move |el_h, core_tx, net_tx| {
+                Client::registered(&locator, &password, &invitation, el_h, core_tx, net_tx)
+            },
+            network_observer,
+        )
+    }
+
+    /// Create a new account
+    fn create_acc_impl<F: 'static + Send, NetObs>(
+        create_client_fn: F,
+        mut network_observer: NetObs,
+    ) -> Result<Self, AuthError>
+    where
+        NetObs: FnMut(Result<NetworkEvent, ()>) + Send + 'static,
+        F: FnOnce(Handle, CoreMsgTx<()>, NetworkTx) -> Result<Client<()>, CoreError>,
+    {
+        let (tx, rx) = sync_channel(0);
 
         let joiner = thread::named("Core Event Loop", move || {
             let el = try_tx!(Core::new(), tx);
@@ -160,17 +180,7 @@ impl Authenticator {
                 .for_each(|_| Ok(()));
             el_h.spawn(net_obs_fut);
 
-            let client = try_tx!(
-                Client::registered(
-                    &locator,
-                    &password,
-                    &invitation,
-                    el_h,
-                    core_tx_clone,
-                    net_tx,
-                ),
-                tx
-            );
+            let client = try_tx!(create_client_fn(el_h, core_tx_clone, net_tx), tx);
 
             let tx2 = tx.clone();
             let core_tx2 = core_tx.clone();
@@ -225,16 +235,32 @@ impl Authenticator {
     pub fn login<S, NetObs>(
         locator: S,
         password: S,
-        mut network_observer: NetObs,
+        network_observer: NetObs,
     ) -> Result<Self, AuthError>
     where
         S: Into<String>,
         NetObs: FnMut(Result<NetworkEvent, ()>) + Send + 'static,
     {
-        let (tx, rx) = sync_channel(0);
 
         let locator = locator.into();
         let password = password.into();
+
+        Self::login_impl(
+            move |el_h, core_tx, net_tx| Client::login(&locator, &password, el_h, core_tx, net_tx),
+            network_observer,
+        )
+    }
+
+    /// Log in to an existing account
+    pub fn login_impl<F: Send + 'static, NetObs>(
+        create_client_fn: F,
+        mut network_observer: NetObs,
+    ) -> Result<Self, AuthError>
+    where
+        F: FnOnce(Handle, CoreMsgTx<()>, NetworkTx) -> Result<Client<()>, CoreError>,
+        NetObs: FnMut(Result<NetworkEvent, ()>) + Send + 'static,
+    {
+        let (tx, rx) = sync_channel(0);
 
         let joiner = thread::named("Core Event Loop", move || {
             let el = try_tx!(Core::new(), tx);
@@ -249,10 +275,7 @@ impl Authenticator {
                 .for_each(|_| Ok(()));
             el_h.spawn(net_obs_fut);
 
-            let client = try_tx!(
-                Client::login(&locator, &password, el_h, core_tx_clone, net_tx),
-                tx
-            );
+            let client = try_tx!(create_client_fn(el_h, core_tx_clone, net_tx), tx);
 
             unwrap!(tx.send(Ok(core_tx)));
 
@@ -265,6 +288,73 @@ impl Authenticator {
             core_tx: Mutex::new(core_tx),
             _core_joiner: joiner,
         })
+    }
+}
+
+#[cfg(all(feature = "use-mock-routing", any(test, feature = "testing")))]
+impl Authenticator {
+    #[allow(unused)]
+    fn login_with_hook<F, S, NetObs>(
+        locator: S,
+        password: S,
+        network_observer: NetObs,
+        routing_wrapper_fn: F,
+    ) -> Result<Self, AuthError>
+    where
+        S: Into<String>,
+        F: Fn(MockRouting) -> MockRouting + Send + 'static,
+        NetObs: FnMut(Result<NetworkEvent, ()>) + Send + 'static,
+    {
+
+        let locator = locator.into();
+        let password = password.into();
+
+        Self::login_impl(
+            move |el_h, core_tx, net_tx| {
+                Client::login_with_hook(
+                    &locator,
+                    &password,
+                    el_h,
+                    core_tx,
+                    net_tx,
+                    routing_wrapper_fn,
+                )
+            },
+            network_observer,
+        )
+    }
+
+    #[allow(unused)]
+    fn create_acc_with_hook<F, S, NetObs>(
+        locator: S,
+        password: S,
+        invitation: S,
+        network_observer: NetObs,
+        routing_wrapper_fn: F,
+    ) -> Result<Self, AuthError>
+    where
+        NetObs: FnMut(Result<NetworkEvent, ()>) + Send + 'static,
+        F: Fn(MockRouting) -> MockRouting + Send + 'static,
+        S: Into<String>,
+    {
+        let locator = locator.into();
+        let password = password.into();
+        let invitation = invitation.into();
+
+        Self::create_acc_impl(
+            move |el_h, core_tx_clone, net_tx| {
+                Client::registered_with_hook(
+                    &locator,
+                    &password,
+                    &invitation,
+                    el_h,
+                    core_tx_clone,
+                    net_tx,
+                    routing_wrapper_fn,
+                )
+            },
+            network_observer,
+        )
     }
 }
